SSF: address review feedback on last poll timestamp
- Clarify that lastPollCompletedAt is stamped after all transmitter-side
work (not at request arrival) and why it stays inside the request
transaction rather than a response-completion hook
- Use Long for lastPollCompletedAt in the admin representation to avoid
the int-seconds 2038 limit for new code
- Make the poll test assert the 60s write-coalescing by advancing the
clock inside the window and expecting an unchanged stamp

+++ ssf/transmitter/src/main/java/org/keycloak/ssf/transmitter/admin/SsfClientStreamRepresentation.java
@@ -43,7 +43,7 @@ public class SsfClientStreamRepresentation {
      * for POLL-delivery streams. Accurate to within
      * {@code SsfActivityTracker.POLL_STAMP_GRANULARITY_SECONDS}.
      */
-    private Integer lastPollCompletedAt;
+    private Long lastPollCompletedAt;
 
     /**
      * Origin / ownership marker — see
@@ -150,11 +150,11 @@ public class SsfClientStreamRepresentation {
         this.lastVerifiedAt = lastVerifiedAt;
     }
 
-    public Integer getLastPollCompletedAt() {
+    public Long getLastPollCompletedAt() {
         return lastPollCompletedAt;
     }
 
-    public void setLastPollCompletedAt(Integer lastPollCompletedAt) {
+    public void setLastPollCompletedAt(Long lastPollCompletedAt) {
         this.lastPollCompletedAt = lastPollCompletedAt;
     }
